Add delete option to every hero listed in heroes.component
- Bind delete button to the delete method in heroes component and limit
propagation of the click event to the button element so that elements
deleted in button element are not selected within the parent list element
- The delete button in the heroes.component requires the hero's id and
calls the delete method of the heroService. Since the heroes.component
is responsible for updating the display, it has to remove the deleted
hero from the hero array in the component
- The delete method of the heroService sends a http delete request to
the InMemory API and a null is returned if the request is successful.
Else, the resulting error will be handled